post access control & toggle like

+++ api/controllers/post.js
@@ -57,6 +57,8 @@ export async function findAll(req, res, next) {
 
 export async function findOne(req, res, next) {
     console.log("POST CONTROLLER FINDONE");
+    
+    if (req.path === "/like") return next();
 
     const { postId } = req.params;
 
@@ -89,13 +91,16 @@ export async function updateOne(req, res, next) {
             = await joi.postEditSchema.validateAsync(req.body);
         const post = {
             postId: req.params.postId,
-            userId: req.app.locals.userId,
+            userId: req.app.locals.user,
             title, content
         }
 
         const result = await Post.updateOne(post);
-        if (result[0] === 0) {
-            throw new Error("UPDATE FAIL");
+        switch (result[0]) {
+            case null:
+                throw new Error("수정 권한이 없습니다");
+            case 0:
+                throw new Error("INTERNAL UPDATE FAILURE");
         }
 
         res.status(200).json({
@@ -111,20 +116,52 @@ export async function updateOne(req, res, next) {
 
 export async function deleteOne(req, res, next) {
     console.log("POST CONTROLLER DELETEONE");
+    const { userId } = req.app.locals.user;
     const { postId } = req.params;
     
-    const result = await Post.deleteOne(postId);
-    if (result === 0) {
-        const error = new Error("DELETE FAIL");
+    try {
+        const result = await Post.deleteOne({ userId, postId });
+        switch (result) {
+            case null:
+                throw new Error("삭제 권한이 없습니다.");
+            case 0:
+                throw new Error("INTERNAL DELETE FAILURE");
+        }
+    
+        res.status(200).json({
+            message: "게시글을 삭제했습니다."
+        });
+        
+    } catch (error) {
+        res.status(400).json({
+            message: error.message
+        });
     }
-
-    res.status(200).json({
-        message: "게시글을 삭제했습니다."
-    });
 }
 
 export async function toggleLike(req, res, next) {
     console.log("POST CONTROLLER TOGGLELIKE");
+
+    const { postId } = req.params;
+    const { userId } = req.app.locals.user;
+
+    try {
+        const result = await Post.toggleLike({ postId, userId });
+        if (result === 1) {
+            res.status(200).json({
+                message: "게시글의 좋아요를 취소했습니다."
+            });
+        } else {
+            res.status(200).json({
+                message: "게시글의 좋아요를 등록했습니다."
+            });
+        }
+        
+    } catch (error) {
+        res.status(400).json({
+            message: error.message
+        });
+    }
 }
 
 export async function likeList(req, res, next) {
